requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-filter==24.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Faker==33.1.0
flake8==7.1.1
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
phonenumbers==8.13.51
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
